backend/requirements.txt
# Core web framework
fastapi==0.135.3
uvicorn==0.43.0

# File handling
python-multipart==0.0.24

# Firebase Admin SDK - for verifying login tokens
firebase-admin==6.6.0

# Environment variables
python-dotenv==1.0.1

# Machine Learning
torch==2.5.1
torchvision==0.20.1
Pillow==11.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.26.0

# Video Processing
opencv-python>=4.8.0
